+++ packages/lib/sdk/src/utils/index.d.ts
@@ -1,3 +1,0 @@
-export * from './index.js';
-export * from './batchUp.js';
-export * from '../lib/sharedPromise.js';