+++ packages/ui/core-components/src/lib/unsorted/ui/Tabs/Tabs.stories.svelte
@@ -8,10 +8,7 @@
 </script>
 
 <script>
-	import { setContext } from 'svelte';
-	import { writable } from 'svelte/store';
 	import { Story } from '@storybook/addon-svelte-csf';
-	import { INPUTS_CONTEXT_KEY } from '@evidence-dev/component-utilities/globalContexts';
 	import { query } from '@evidence-dev/universal-sql/client-duckdb';
 	import { Query } from '@evidence-dev/sdk/usql';
 	import Tab from './Tab.svelte';
@@ -19,8 +16,8 @@
 	import { within, userEvent } from '@storybook/testing-library';
 	import { expect } from '@storybook/jest';
 
-	const inputStore = writable({});
-	setContext(INPUTS_CONTEXT_KEY, inputStore);
+	import { getInputContext } from '@evidence-dev/sdk/utils/svelte';
+	const inputStore = getInputContext();
 
 	$: data = Query.create(
 		`
@@ -78,5 +75,26 @@
 		<Tab label="Tab 1" id="tab1">Tab 1 content</Tab>
 		<Tab label="Tab 2" id="tab2">Tab 2 content</Tab>
 		<Tab label="Tab 3" id="tab3">Tab 3 content</Tab>
+	</Tabs>
+</Story>
+
+<Story name="Tabs w/ id">
+	<Tabs id>
+		<Tab label="Tab 1" id="tab1">Tab 1 content</Tab>
+		<Tab label="Tab 2" id="tab2">Tab 2 content</Tab>
+		<Tab label="Tab 3" id="tab3">Tab 3 content</Tab>
+	</Tabs>
+</Story>
+
+<Story name="Tabs w/ multiple id">
+	<Tabs id="first-set">
+		<Tab label="Tab 1" id="tab1">Tab 1 content</Tab>
+		<Tab label="Tab 2" id="tab2">Tab 2 content</Tab>
+		<Tab label="Tab 3" id="tab3">Tab 3 content</Tab>
+	</Tabs>
+	<Tabs id="second-set">
+		<Tab label="Tab 4" id="tab4">Tab 4 content</Tab>
+		<Tab label="Tab 4" id="tab5">Tab 5 content</Tab>
+		<Tab label="Tab 6" id="tab6">Tab 6 content</Tab>
 	</Tabs>
 </Story>